+++ burstphoto/io_dng/dng_sdk_wrapper.h
@@ -11,6 +11,7 @@ extern "C" {
     void initialize_xmp_sdk();
     void terminate_xmp_sdk();
 
+
     // function to read a dng image and store its pixel values
     int read_dng_from_disk(const char* in_path, void** pixel_bytes_pointer, int* width, int* height, int* mosaic_pattern_width, int* white_level, int* black_level, int* masked_areas, int* exposure_bias, float* ISO_exposure_time, float* color_factor_r, float* color_factor_g, float* color_factor_b);
